\name{bshazard-package}
\alias{bshazard-package}
\docType{package}
\title{
Nonparametric Smoothing of the Hazard Function
}
\description{
The function estimates the hazard function non parametrically from a
survival object (possibly adjusted for covariates). The smoothed
estimate is based on B-splines from the perspective of generalized
linear mixed models. Left truncated and right censoring data are
allowed.
}
\details{
\tabular{ll}{
Package: \tab bshazard\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-02-01\cr
License: \tab GPL-2\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Paola Rebora, Agus Salim, Marie Reilly
Maintainer: Paola Rebora <paola.rebora@unimib.it>
}
\references{
Rebora P, Salim A, Reilly M bshazard: A Flexible Tool for Nonparametric
Smoothing of the Hazard Function

Lee Y, Nelder JA, Pawitan Y (2006). Generalized Linear Models with Random Effects: Unified
Analysis via H-likelihood, volume 106. Chapman & Hall/CRC.

Pawitan Y (2001). In All Likelihood: Statistical Modelling and Inference Using Likelihood.
Oxford University Press
}
\keyword{ hazard rate }
